apps/briefs/requirements.txt
openai
json-repair
google-genai
gliclass
rapidfuzz
transformers>=4.48.0
retry
ipywidgets
widgetsnbextension
pandas-profiling
readtime
optuna
bertopic
supabase
fastapi
uvicorn
psycopg2-binary
pandas
torch
hdbscan
umap-learn
scikit-learn
matplotlib
seaborn
tiktoken
python-dotenv
